+++ packages/nesd/lib/ui/settings/controls/controls_settings.dart
@@ -12,6 +12,7 @@ import 'package:nesd/ui/settings/controls/binding_tile.dart';
 import 'package:nesd/ui/settings/controls/reset_bindings_button.dart';
 import 'package:nesd/ui/settings/controls/show_touch_controls_switch.dart';
 import 'package:nesd/ui/settings/controls/touch_editor_button.dart';
+import 'package:nesd/ui/settings/controls/turbo_speed_selector.dart';
 import 'package:nesd/ui/settings/settings.dart';
 import 'package:nesd/ui/settings/settings_tab.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
@@ -30,6 +31,7 @@ class ControlsSettings extends StatelessWidget {
           children: [
             const ResetBindingsButton(),
             const ShowTouchControlsSwitch(),
+            const TurboSpeedSelector(),
             const TouchEditorButton(),
             const ProfileSelectionHeader(),
             for (final action in allActions)